n64: improve address exceptions with correct COP0 status
Fix several tests in n64-systemtest ROM.

+++ ares/n64/cpu/interpreter-ipu.cpp
@@ -367,16 +367,16 @@ auto CPU::JR(cr64& rs) -> void {
 }
 
 auto CPU::LB(r64& rt, cr64& rs, s16 imm) -> void {
-  if(auto data = read<Byte>(rs.u32 + imm)) rt.u64 = s8(*data);
+  if(auto data = read<Byte>(rs.u64 + imm)) rt.u64 = s8(*data);
 }
 
 auto CPU::LBU(r64& rt, cr64& rs, s16 imm) -> void {
-  if(auto data = read<Byte>(rs.u32 + imm)) rt.u64 = u8(*data);
+  if(auto data = read<Byte>(rs.u64 + imm)) rt.u64 = u8(*data);
 }
 
 auto CPU::LD(r64& rt, cr64& rs, s16 imm) -> void {
   if(!context.kernelMode() && context.bits == 32) return exception.reservedInstruction();
-  if(auto data = read<Dual>(rs.u32 + imm)) rt.u64 = *data;
+  if(auto data = read<Dual>(rs.u64 + imm)) rt.u64 = *data;
 }
 
 auto CPU::LDL(r64& rt, cr64& rs, s16 imm) -> void {
@@ -560,15 +560,15 @@ auto CPU::LDR(r64& rt, cr64& rs, s16 imm) -> void {
 }
 
 auto CPU::LH(r64& rt, cr64& rs, s16 imm) -> void {
-  if(auto data = read<Half>(rs.u32 + imm)) rt.u64 = s16(*data);
+  if(auto data = read<Half>(rs.u64 + imm)) rt.u64 = s16(*data);
 }
 
 auto CPU::LHU(r64& rt, cr64& rs, s16 imm) -> void {
-  if(auto data = read<Half>(rs.u32 + imm)) rt.u64 = u16(*data);
+  if(auto data = read<Half>(rs.u64 + imm)) rt.u64 = u16(*data);
 }
 
 auto CPU::LL(r64& rt, cr64& rs, s16 imm) -> void {
-  if(auto data = read<Word>(rs.u32 + imm)) {
+  if(auto data = read<Word>(rs.u64 + imm)) {
     rt.u64 = s32(*data);
     scc.ll = tlb.physicalAddress >> 4;
     scc.llbit = 1;
@@ -577,7 +577,7 @@ auto CPU::LL(r64& rt, cr64& rs, s16 imm) -> void {
 
 auto CPU::LLD(r64& rt, cr64& rs, s16 imm) -> void {
   if(!context.kernelMode() && context.bits == 32) return exception.reservedInstruction();
-  if(auto data = read<Dual>(rs.u32 + imm)) {
+  if(auto data = read<Dual>(rs.u64 + imm)) {
     rt.u64 = *data;
     scc.ll = tlb.physicalAddress >> 4;
     scc.llbit = 1;
@@ -589,7 +589,7 @@ auto CPU::LUI(r64& rt, u16 imm) -> void {
 }
 
 auto CPU::LW(r64& rt, cr64& rs, s16 imm) -> void {
-  if(auto data = read<Word>(rs.u32 + imm)) rt.u64 = s32(*data);
+  if(auto data = read<Word>(rs.u64 + imm)) rt.u64 = s32(*data);
 }
 
 auto CPU::LWL(r64& rt, cr64& rs, s16 imm) -> void {
@@ -703,7 +703,7 @@ auto CPU::LWR(r64& rt, cr64& rs, s16 imm) -> void {
 }
 
 auto CPU::LWU(r64& rt, cr64& rs, s16 imm) -> void {
-  if(auto data = read<Word>(rs.u32 + imm)) rt.u64 = u32(*data);
+  if(auto data = read<Word>(rs.u64 + imm)) rt.u64 = u32(*data);
 }
 
 auto CPU::MFHI(r64& rd) -> void {
@@ -755,7 +755,7 @@ auto CPU::SB(cr64& rt, cr64& rs, s16 imm) -> void {
 auto CPU::SC(r64& rt, cr64& rs, s16 imm) -> void {
   if(scc.llbit) {
     scc.llbit = 0;
-    rt.u64 = write<Word>(rs.u32 + imm, rt.u32);
+    rt.u64 = write<Word>(rs.u64 + imm, rt.u32);
   } else {
     rt.u64 = 0;
   }
@@ -765,7 +765,7 @@ auto CPU::SCD(r64& rt, cr64& rs, s16 imm) -> void {
   if(!context.kernelMode() && context.bits == 32) return exception.reservedInstruction();
   if(scc.llbit) {
     scc.llbit = 0;
-    rt.u64 = write<Dual>(rs.u32 + imm, rt.u64);
+    rt.u64 = write<Dual>(rs.u64 + imm, rt.u64);
   } else {
     rt.u64 = 0;
   }
@@ -773,7 +773,7 @@ auto CPU::SCD(r64& rt, cr64& rs, s16 imm) -> void {
 
 auto CPU::SD(cr64& rt, cr64& rs, s16 imm) -> void {
   if(!context.kernelMode() && context.bits == 32) return exception.reservedInstruction();
-  write<Dual>(rs.u32 + imm, rt.u64);
+  write<Dual>(rs.u64 + imm, rt.u64);
 }
 
 auto CPU::SDL(cr64& rt, cr64& rs, s16 imm) -> void {
@@ -921,7 +921,7 @@ auto CPU::SDR(cr64& rt, cr64& rs, s16 imm) -> void {
 }
 
 auto CPU::SH(cr64& rt, cr64& rs, s16 imm) -> void {
-  write<Half>(rs.u32 + imm, rt.u32);
+  write<Half>(rs.u64 + imm, rt.u32);
 }
 
 auto CPU::SLL(r64& rd, cr64& rt, u8 sa) -> void {
@@ -974,7 +974,7 @@ auto CPU::SUBU(r64& rd, cr64& rs, cr64& rt) -> void {
 }
 
 auto CPU::SW(cr64& rt, cr64& rs, s16 imm) -> void {
-  write<Word>(rs.u32 + imm, rt.u32);
+  write<Word>(rs.u64 + imm, rt.u32);
 }
 
 auto CPU::SWL(cr64& rt, cr64& rs, s16 imm) -> void {
